// workspaces/scorecard/plugins/scorecard-backend-module-code-coverage/src/clients/types.ts
/*
 * Copyright Red Hat, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

/**
 * Represents a single coverage aggregate section (line or branch).
 */
export interface CoverageAggregate {
  available: number;
  covered: number;
  missed: number;
  percentage: number;
}

/**
 * The response from the code-coverage-backend report API.
 */
export interface CodeCoverageReport {
  aggregate: {
    line: CoverageAggregate;
    branch: CoverageAggregate;
  };
  entity: {
    kind: string;
    name: string;
    namespace: string;
  };
  files: Array<{
    branchHits: Record<string, number>;
    filename: string;
    lineHits: Record<string, number>;
  }>;
}